Finish adding fences and change camera position
Finished enclosing the school with fences and added collision boxes to them. There are currently four directions that enemies approach from, but we may want to modify this to be more.

The camera position has also been changed to something I think better fits what we're going for. Let me know if this is a good or bad change.