Add animations and character model!
We finally have a decent character model! And it animates! This was just a quick test to port some animations in from the asset store, and it worked! It's not the most pretty, but it works. There are currently a bunch of unused animations in here which we can pick through. The animation controller is a little difficult to understand. It came from this asset package: https://assetstore.unity.com/packages/3d/animations/rpg-character-mecanim-animation-pack-free-65284.