You can now hide children
Children can now be hidden in bushes or in the tent (or any object that is on the "Stash" layer and has a collider). Children stay in hiding for a set amount of time, but get impatient and start wandering after a while. Children are safe while in hiding, giving you a break from taking care of them.