Add alert arrow and fix enemy attacks
There is now an alert arrow that points to children when they are picked up by enemies. It goes away when either the children are taken out of map or the enemies are killed.

I also fixed enemy attacks and killed a couple bugs that were generating a lot of console spam.