Add high score menu
I added a high score menu to the main menu. It saves and loads scores using the PlayerPrefs class. In order to store the entire list of them, I serialize a list of scores in json format to add to the PlayerPrefs. Unfortunately, this involved the creation of a couple more classes than I would like, but I did it this way to make it easier to implement further. In game scores are not saved to the list yet, I still need to add that.