Finish intro scene and change main menu background
Made some final touches to the intro comic and added a more fitting background to the main menu. I would have liked to finish the comic more, but I don't have enough time to bring them to the level of detail I would really want, so I'm going to leave it here for now.

Fixes #79 fixes #92